get rid of a warning
git-svn-id: file:///home/shish/svn/shimmie2/trunk@38 7f39781d-f577-437e-ae19-be835c7a54ca

+++ ext/regen_thumb.ext.php
@@ -37,7 +37,7 @@ class RegenThumb extends Extension {
 
 		switch($program) {
 			case 'convert':
-				unlink($f_thumb);
+				if(file_exists($f_thumb)) unlink($f_thumb);
 				exec("convert $f_image -geometry {$w}x{$h} -quality {$q} $f_thumb");
 				break;
 			case 'gd':
